refactor(server): extract council route domain (#588)
Move the council seating board's 7 routes + private helpers and the
background watch loop out of server.py into routes/council.py as
CouncilRoutesMixin. Pure relocation — no route added or dropped.

Closes #588

Built by dotdev.dev
